… src/cmd/internal/objabi/reloctype_string.go