In some runtime functions, there are conditional branches that lead to absolute branches. For example:
```
// func memhash(p unsafe.Pointer, h, size uintptr) uintptr
TEXT runtime·memhash<ABIInternal>(SB),NOSPLIT|NOFRAME,$0-32
	MOVB	runtime·useAeshash(SB), R10
	CBZ	R10, noaes
	B	aeshashbody<>(SB)
noaes:
	B	runtime·memhashFallback<ABIInternal>(SB)
```
This patch allows direct conditional branches. This affects hash function written in runtime/asm_arm64.s.
Results of running runtime hash benchmarks:
```
goos: linux
goarch: arm64
pkg: runtime
                       │ master_noaes_runtime_hash.txt │ arm64-condbr19_noaes_runtime_hash.txt │
                       │            sec/op             │     sec/op      vs base               │
Hash5-4                                    9.628n ± 0%      8.858n ± 0%  -8.00% (p=0.000 n=25)
Hash16-4                                   9.627n ± 0%      9.627n ± 0%       ~ (p=0.117 n=25)
Hash64-4                                   15.03n ± 0%      14.25n ± 0%  -5.19% (p=0.000 n=25)
Hash1024-4                                 90.19n ± 0%      88.19n ± 0%  -2.22% (p=0.000 n=25)
Hash65536-4                                5.058µ ± 0%      5.192µ ± 0%  +2.65% (p=0.000 n=25)
HashStringSpeed-4                          19.00n ± 3%      18.46n ± 4%       ~ (p=0.096 n=25)
HashBytesSpeed-4                           65.26n ± 0%      64.36n ± 0%  -1.38% (p=0.000 n=25)
HashInt32Speed-4                           12.46n ± 0%      12.50n ± 1%       ~ (p=0.829 n=25)
HashInt64Speed-4                           13.01n ± 3%      12.47n ± 3%  -4.15% (p=0.000 n=25)
HashStringArraySpeed-4                     43.45n ± 2%      41.55n ± 3%  -4.37% (p=0.000 n=25)
FastrandHashiter-4                         11.72n ± 1%      11.81n ± 1%       ~ (p=0.079 n=25)
geomean                                    33.77n           33.00n       -2.26%

            │ master_noaes_runtime_hash.txt │ arm64-condbr19_noaes_runtime_hash.txt │
            │              B/s              │      B/s        vs base               │
Hash5-4                        495.3Mi ± 0%     538.3Mi ± 0%  +8.69% (p=0.000 n=25)
Hash16-4                       1.548Gi ± 0%     1.548Gi ± 0%  +0.00% (p=0.025 n=25)
Hash64-4                       3.966Gi ± 0%     4.183Gi ± 0%  +5.47% (p=0.000 n=25)
Hash1024-4                     10.57Gi ± 0%     10.81Gi ± 0%  +2.27% (p=0.000 n=25)
Hash65536-4                    12.07Gi ± 0%     11.76Gi ± 0%  -2.57% (p=0.000 n=25)
geomean                        3.279Gi          3.367Gi       +2.70%
```
These results are obtained with cpu.ARM64.HasAES == false

Change-Id: If282267b9b2dcf474516cb33bfdbda4ee35bb8fa